// src/app/models/product.model.ts
// Generated by https://quicktype.io

export interface Product {
  id:          number;
  title:       string;
  price:       number;
  description: string;
  category:    Category;
  images:      string[];
  taxes?:number;
}

export interface ProductDataTrnsferObject extends Omit<Product, 'id' | 'category'>{
  categoryId: number;
}

export interface ProductUpdateDTO extends Partial<ProductDataTrnsferObject>{}

export interface Category {
  id:      number;
  name:    Name;
  typeImg: TypeImg;
}

export enum Name {
  Clothes = "Clothes",
  Electronics = "Electronics",
  Furniture = "Furniture",
  Others = "Others",
  Toys = "Toys",
}

export enum TypeImg {
  Animals = "animals",
  Any = "any",
  Arch = "arch",
  People = "people",
  Tech = "tech",
}